# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Search Platform API",
        "description": "API Documentation for Job Search Platform Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/cvs": {
            "get": {
                "tags": [
                    "CVs"
                ],
                "summary": "Get my CVs",
                "description": "Returns a list of CVs created by the authenticated job seeker.",
                "operationId": "getCVs",
                "responses": {
                    "200": {
                        "description": "List of CVs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CVs"
                ],
                "summary": "Create a new CV",
                "description": "Create a new CV.",
                "operationId": "createCV",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Full Stack Developer CV"
                                    },
                                    "personal_summary": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CV created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cvs/{id}": {
            "get": {
                "tags": [
                    "CVs"
                ],
                "summary": "Get CV details",
                "description": "Returns full details of a specific CV, including skills, education, experience, etc.",
                "operationId": "getCV",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CV details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CVs"
                ],
                "summary": "Update CV details",
                "description": "Update a CV.",
                "operationId": "updateCV",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "personal_summary": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CV updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CVs"
                ],
                "summary": "Delete a CV",
                "description": "Delete a CV.",
                "operationId": "deleteCV",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CV deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cvs/{cvId}/skills": {
            "post": {
                "tags": [
                    "CVs",
                    "Skills"
                ],
                "summary": "Add skill to CV",
                "description": "Add a skill to CV.",
                "operationId": "addCVSkill",
                "parameters": [
                    {
                        "name": "cvId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skill_id"
                                ],
                                "properties": {
                                    "skill_id": {
                                        "type": "integer"
                                    },
                                    "skill_level": {
                                        "type": "string",
                                        "example": "Intermediate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill added to CV"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cvs/{cvId}/skills/{skillId}": {
            "delete": {
                "tags": [
                    "CVs",
                    "Skills"
                ],
                "summary": "Remove skill from CV",
                "description": "Remove a skill from CV.",
                "operationId": "removeCVSkill",
                "parameters": [
                    {
                        "name": "cvId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "skillId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill removed from CV"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cvs/{cvId}/education": {
            "post": {
                "tags": [
                    "CVs",
                    "Education"
                ],
                "summary": "Add education to CV",
                "description": "Add education to CV.",
                "operationId": "addCVEducation",
                "parameters": [
                    {
                        "name": "cvId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "institution",
                                    "degree_name"
                                ],
                                "properties": {
                                    "institution": {
                                        "type": "string"
                                    },
                                    "degree_name": {
                                        "type": "string"
                                    },
                                    "major": {
                                        "type": "string"
                                    },
                                    "graduation_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Education added to CV"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cvs/{cvId}/education/{educationId}": {
            "delete": {
                "tags": [
                    "CVs",
                    "Education"
                ],
                "summary": "Remove education from CV",
                "description": "Remove education from CV.",
                "operationId": "removeCVEducation",
                "parameters": [
                    {
                        "name": "cvId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "educationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Education removed from CV"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}